requirements.txt
torch==2.2.0
loguru==0.7.2
flask==2.2.5
python-dotenv
web3
ezkl
onnx
solc-select